Move reduction tests to reduction package
